% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_minicutter.R
\name{calculate_usage_ratios}
\alias{calculate_usage_ratios}
\title{Helper function to get junction usage ratios
calculate_usage_ratios
This function takes}
\usage{
calculate_usage_ratios(juncs)
}
\arguments{
\item{juncs}{A list of integers representing the sizes of each cluster}
}
\value{
A plot
}
\description{
Helper function to get junction usage ratios
calculate_usage_ratios
This function takes
}
\examples{
calculate_usage_ratios(juncs)

}
