% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_coords.R
\name{isoviz_coords}
\alias{isoviz_coords}
\title{A function for obtaining all necessary coordinates for leafcutter integration
and plotting isoforms given a gtf file or psl file from long read sequencing}
\usage{
isoviz_coords(file_path, gene_trans, count_file = "", min_count = 5)
}
\arguments{
\item{file_path}{Gencode PSL file! (can be custom PSL file though)}

\item{gene_trans}{Dataframe with ensembl gene and transcript ids and names}

\item{count_file}{User can provide a read count file per transcript from a long read experiment}

\item{min_count}{Require a specific number of reads per transcript. Only works if providing count file. Default is 5.}

\item{input_type}{Either 'psl' or 'gtf' for input file type. Default is psl.}
}
\value{
list of exon and intron coordinates in that order.
}
\description{
Inputs gencode gtf or psl file, outputs exon coords (for plotting) and
intron coords (for leafcutter integration) with ENSG/ENST names plus
meaningful names (ex RBFOX2)
}
\examples{
# Example with psl file and pre-loaded gene to transcript conversions
psl_file <- system.file("data", "gencode.v41.basic.annotation.psl", package="isoviz")
gene_trans <- system.file("data", "gencode_v41_gene_transcript_convert.txt", package="isoviz")

results <- isoviz_coords(psl_file, gene_trans, input_type = "psl")
print(head(results)[1]) #exons coords
print(head(results)[2]) #introns coords

}
