% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_get_gene_transcript_conversion.R
\name{isoviz_get_gene_transcript_conversion}
\alias{isoviz_get_gene_transcript_conversion}
\title{A function for converting gtf to psl}
\usage{
isoviz_get_gene_transcript_conversion(
  gtf_file,
  output_file = "gene_transcript_convert.txt"
)
}
\arguments{
\item{output_file}{Name for the PSL file to be outputted. Default is "converted_gtf.psl"}

\item{gtf_file_path}{User provided gtf file}
}
\value{
table with gene ids, trancsript ids, names...
}
\description{
Converts a gtf to psl, depending on the output filename extension;
gtf exons need to be grouped by transcript and sorted by coordinate w/in a transcript
this function is adapted from Flair https://github.com/BrooksLabUCSC/flair
}
\examples{
# Example with gtf file 

}
