% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_get_guide_predictions.R
\name{isoviz_get_guide_predictions}
\alias{isoviz_get_guide_predictions}
\title{A function for obtaining junction-level guide RNA efficiency predictions for a specific gene or junction}
\usage{
isoviz_get_guide_predictions(
  gene = "ENSG00000100320",
  cell_type = "hESC",
  guides_per_junction = 8,
  include_specific_junctions = c(),
  leafcutter_input,
  output_format = "dataframe"
)
}
\arguments{
\item{gene}{ensembl ID for now}

\item{cell_type}{cell type of interest, default=hESC}

\item{guides_per_junction}{the number of highest ranking guides to print out, minimum of 1 and maximum of 8, default is 8}

\item{include_specific_junctions}{provide list of junctions for guide RNA prediction}

\item{leafcutter_input}{output from running isoviz_minicutter}

\item{output_format}{set to dataframe to obtain a regular dataframe object, to obtain a nice table, choose "nice_table"}
}
\value{
table
}
\description{
A function for obtaining junction-level guide RNA efficiency predictions for a specific gene or junction
}
\examples{
isoviz_get_guide_predictions()
}
