% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_gtf_to_psl.R
\name{isoviz_gtf_to_psl}
\alias{isoviz_gtf_to_psl}
\title{A function for converting gtf to psl}
\usage{
isoviz_gtf_to_psl(
  gtf_file_path,
  psl_output_file = "converted_gtf.psl",
  gtf_source = "gencode",
  filter_CDS = "no",
  chrom_sizes = NULL
)
}
\arguments{
\item{gtf_file_path}{User provided gtf file}

\item{psl_output_file}{Name for the PSL file to be outputted. Default is "converted_gtf.psl"}

\item{gtf_source}{"gencode" is default otherwise enter "nongencode"}

\item{filter_CDS}{if "yes" then will only include CDS exons in the PSL output file}

\item{chrom_sizes}{Chrom sizes file for psl conversion, recommended}
}
\value{
psl formatted file
}
\description{
Converts a gtf to psl, depending on the output filename extension;
gtf exons need to be grouped by transcript and sorted by coordinate w/in a transcript
this function is adapted from Flair https://github.com/BrooksLabUCSC/flair
}
\examples{
# Example with gtf file 

}
