% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_junction_to_transcript_summary.R
\name{isoviz_junction_to_transcript_summary}
\alias{isoviz_junction_to_transcript_summary}
\title{A function for summarizing junction and transcript level information from GENCODE annotated transcripts or long read sequencing transcripts.
This is especially useful to identify unique junctions that are specific to a single transcript.}
\usage{
isoviz_junction_to_transcript_summary(
  iso_intron_data,
  type = "gencode",
  plot_path = ""
)
}
\arguments{
\item{iso_intron_data}{This is output from isoviz_coords.}

\item{type}{Either 'gencode' or 'longread' depending on the initial source of transcript information. Default is gencode.}

\item{plot_path}{This is optional, but if you would like a summary plot, give the path and name of the output. Include proper extension such as '.pdf' or .'png'. Default is no plot.}
}
\value{
table summarizing junction and transcript level information.
}
\description{
Inputs gencode intron rda file or long read intron rda file (output from isoviz_coords), outputs summary table
and summary plot
}
