% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_minicutter.R
\name{isoviz_minicutter}
\alias{isoviz_minicutter}
\title{isoviz_minicutter function
A function for obtaining leafcutter intron clusters}
\usage{
isoviz_minicutter(
  juncs_file,
  min_junction_reads = 1,
  plot_summary = TRUE,
  min_usage_ratio = 0.001
)
}
\arguments{
\item{juncs_file}{Gencode gtf or psl file}
}
\value{
intron clusters
}
\description{
Inputs junction coordinates from regtools
}
\examples{
isoviz_minicutter()
}
