% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_map_junctions.R,
%   R/isoviz_plot_juncs_to_iso.R
\name{isoviz_plot_juncs_to_iso}
\alias{isoviz_plot_juncs_to_iso}
\title{A function for mapping junctions to transcript isoforms}
\usage{
isoviz_plot_juncs_to_iso(
  gene = "ENSG00000099875",
  cell_type = "hESC",
  junction_usage = 5,
  include_common_juncs = FALSE,
  filter_expressed_isoforms = TRUE,
  filter_all_100percent = FALSE
)
}
\arguments{
\item{gene}{ensembl ID for now}

\item{cell_type}{cell type of interest, default=hESC}

\item{junction_usage}{minimum junction usage, default is 5}

\item{include_common_juncs}{show junctions common to all isoforms?}

\item{filter_expressed_isoforms}{include only isoforms that are detected}

\item{filter_all_100percent}{filter junctions that are always included (is this right?)}

\item{intron_coords}{description}

\item{leafcutter_input}{output from running isoviz_minicutter}
}
\value{
mapping

plot
}
\description{
A function for mapping junctions to transcript isoforms

A function for obtaining a comprehensive visualization of gene transcript
isoforms with junction coverage across intron clusters information from RNA-seq
and leafcutter clustering
}
\examples{
isoviz_plot_juncs_to_iso()
isoviz_plot_juncs_to_iso()
}
