% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_rescale_introns.R
\name{isoviz_rescale_introns}
\alias{isoviz_rescale_introns}
\title{A function for reducing the length of introns when visualizing isoforms
through re-scaling all coordinates in the gene.}
\usage{
isoviz_rescale_introns(introns, exons, width_rescale = 50)
}
\arguments{
\item{introns}{#description}

\item{exons}{#description}

\item{width_rescale}{#description}

\item{transcripts_group}{Column name containing transcript name}
}
\value{
'data.table()' containing the re-scaled coordinates of 'introns'
and 'exons' for all transcripts in input
}
\description{
Re-scaling intron coordinates while perserving exon alignment by only
impacting the intron width and not the exons. These re-scaled coordinates are
meant only for plotting!
}
\examples{
isoviz_rescale_introns()
}
