% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_minicutter.R
\name{leafcutter_one_step}
\alias{leafcutter_one_step}
\title{leafcutter_one_step function
A function for obtaining leafcutter intron clusters}
\usage{
leafcutter_one_step(juncs)
}
\arguments{
\item{juncs_file}{Gencode gtf or psl file}
}
\value{
intron clusters
}
\description{
Inputs junction coordinates from regtools
}
\examples{
leafcutter_one_step()
}
