% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minicutter.R
\name{minicutter}
\alias{minicutter}
\alias{leafcutter_one_step}
\title{A function for obtaining leafcutter intron clusters}
\usage{
leafcutter_one_step(juncs)
}
\arguments{
\item{juncs}{Gencode gtf or psl file}
}
\value{
intron clusters
}
\description{
Inputs junction coordinates from regtools
}
\examples{
minicutter()
}
