% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_minicutter.R
\name{plot_cluster_sizes}
\alias{plot_cluster_sizes}
\title{Helper function to plot cluster sizes
plot_cluster_sizes
This function takes a list of cluster sizes and plots a bar chart of the sizes.}
\usage{
plot_cluster_sizes(juncs)
}
\arguments{
\item{juncs}{A list of integers representing the sizes of each cluster}
}
\value{
A plot
}
\description{
Helper function to plot cluster sizes
plot_cluster_sizes
This function takes a list of cluster sizes and plots a bar chart of the sizes.
}
\examples{
plot_cluster_sizes(juncs)

}
