% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_isoforms.R
\name{plot_isoforms}
\alias{plot_isoforms}
\title{A function for plotting transcript isoforms for a specific gene}
\arguments{
\item{gene}{Input your gene of interest using ENSG for now. Defaults to "ENSG00000100320" (RBFOX2)}
}
\value{
A plot.
}
\description{
This function allows you to visualize all annotated isoforms for a given gene.
}
\examples{
plot_isoforms("ENSG00000100320")
}
