% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_coords.R
\name{isoviz_coords}
\alias{isoviz_coords}
\title{A function for obtaining all necessary coordinates for leafcutter integration
and plotting isoforms given a gtf file or psl file from long read sequencing}
\usage{
isoviz_coords(file_path, gene_trans, input_type = "psl")
}
\arguments{
\item{file_path}{Gencode gtf or psl file}

\item{gene_trans}{Dataframe with ensembl gene and transcript ids and names}
}
\value{
list of exon and intron coordinates in that order.
}
\description{
Inputs gencode gtf or psl file, outputs exon coords (for plotting) and
intron coords (for leafcutter integration) with ENSG/ENST names plus
meaningful names (ex RBFOX2)
}
\examples{
isoviz_coords()
}
