% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_map_junctions.R
\name{isoviz_map_junctions}
\alias{isoviz_map_junctions}
\title{A function for mapping junctions to transcript isoforms}
\arguments{
\item{gene_intron_coords}{obtained intron coordinates for gene of interest}

\item{leafcutter_input}{output from running isoviz_minicutter}

\item{cell_type}{cell type of interest, default=hESC}

\item{gencode_intron_all_data}{expanded annotations of introns/transcripts}
}
\value{
mapping
}
\description{
A function for mapping junctions to transcript isoforms
}
\examples{
isoviz_map_junctions()
}
