% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_plot_juncs_to_iso.R,
%   R/map_all_junctions.R
\name{isoviz_plot_juncs_to_iso}
\alias{isoviz_plot_juncs_to_iso}
\alias{map_all_junctions}
\title{A function for obtaining a comprehensive visualization of gene transcript
isoforms with junction coverage across intron clusters information from RNA-seq
and leafcutter clustering}
\usage{
isoviz_plot_juncs_to_iso(
  gene = "ENSG00000099875",
  cell_type = "hESC",
  junction_usage = 5,
  include_common_juncs = FALSE,
  filter_expressed_isoforms = TRUE,
  filter_all_100percent = FALSE
)

map_all_junctions(
  gene = "ENSG00000100320",
  cell_type = "hESC",
  intron_coords,
  leafcutter_input
)
}
\arguments{
\item{gene}{ensembl ID for now}

\item{cell_type}{cell type of interest, default=hESC}

\item{junction_usage}{minimum junction usage, default is 5}

\item{include_common_juncs}{show junctions common to all isoforms?}

\item{filter_expressed_isoforms}{include only isoforms that are detected}

\item{filter_all_100percent}{filter junctions that are always included (is this right?)}
}
\value{
plot

plot
}
\description{
A function for obtaining a comprehensive visualization of gene transcript
isoforms with junction coverage across intron clusters information from RNA-seq
and leafcutter clustering

A function for mapping junctions to transcript isoforms
}
\examples{
isoviz_plot_juncs_to_iso()
isoviz_plot_juncs_to_iso()
}
