% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_plot_juncs_to_iso.R
\name{isoviz_plot_juncs_to_iso}
\alias{isoviz_plot_juncs_to_iso}
\title{A function for obtaining a comprehensive visualization of gene transcript
isoforms with junction coverage across intron clusters information from RNA-seq
and leafcutter clustering}
\usage{
isoviz_plot_juncs_to_iso(
  mapped_junctions,
  gene_data,
  gene_introns,
  cell_type = "hESC",
  junc_usage = 5,
  intron_scale = "no",
  intron_scale_width = 10,
  include_specific_isoforms = c(),
  include_specific_junctions = c(),
  plot_transcripts_only = FALSE
)
}
\arguments{
\item{mapped_junctions}{output from running isoviz_map_junctions}

\item{cell_type}{cell type of interest, default=hESC}

\item{include_specific_isoforms}{provide list of isoforms to plot in first panel, default is to plot all}

\item{include_specific_junctions}{provide list of junctions to plot in second panel}

\item{plot_transcripts_only}{plot the top plot only showing transcript structure, default is FALSE}

\item{junction_usage}{minimum junction usage as a percentage, default is 5}
}
\value{
plot
}
\description{
A function for obtaining a comprehensive visualization of gene transcript
isoforms with junction coverage across intron clusters information from RNA-seq
and leafcutter clustering
}
\examples{
isoviz_plot_juncs_to_iso()
}
